generate starting stars
initial attempt at creating starting locations for players.
Each player will have 6 stars to start. The first star is chosen
randomly. The remaining stars are chosen by picking an existing star a
player has and then randomly picking a distance. Then randomly pick a
point on this circle as the location for the new star. Check to see if
the new star location is too close to any existing star and repeat the
process if it is

Closes #2