Carrier waypointing
From the carrier menu, the user is able to edit waypoint. When in the
edit waypoint mode, the carriers current destination and all other
destinations within the players hyperspace range are marked as possible
waypoint locations. When any of these stars is clicked, it is added to
the carriers waypoint list and now becomes the new current destination
and the process repeats. Currently this happens only on the client side
and nothing is saved to the server so at the start of the new tick all
waypoints are reset.

Closes #13